refactor: split home viewmodel support responsibilities
