+++ scripts/quest_stage3_ux_reminder_unit_check.swift
@@ -22,6 +22,9 @@ func loadMany(_ relativePaths: [String]) -> String {
 let homeView = load("dogArea/Views/HomeView/HomeView.swift")
 let homeViewModel = loadMany([
     "dogArea/Views/HomeView/HomeViewModel.swift",
+    "dogArea/Views/HomeView/HomeViewModelSupport/HomeViewModel+SessionLifecycle.swift",
+    "dogArea/Views/HomeView/HomeViewModelSupport/HomeViewModel+AreaProgress.swift",
+    "dogArea/Views/HomeView/HomeViewModelSupport/HomeViewModel+IndoorMissionFlow.swift",
     "dogArea/Source/Domain/Home/Models/HomeMissionModels.swift",
     "dogArea/Source/Domain/Home/Services/HomeQuestReminderSupport.swift",
     "dogArea/Source/Domain/Home/Stores/IndoorMissionStore.swift",
